Refactor tests of view helpers
